:tractor: Move processor to its own package
