+++ cmd/glaze/cmds/markdown.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"github.com/go-go-golems/glazed/pkg/cli"
 	"github.com/go-go-golems/glazed/pkg/cmds"
+	"github.com/go-go-golems/glazed/pkg/processor"
 	"github.com/spf13/cobra"
 	"github.com/yuin/goldmark"
 	"github.com/yuin/goldmark/ast"
@@ -184,7 +185,7 @@ var parseCmd = &cobra.Command{
 //		    }
 //
 // ]
-func splitByHeading(md goldmark.Markdown, s []byte, gp *cmds.GlazeProcessor) error {
+func splitByHeading(md goldmark.Markdown, s []byte, gp *processor.GlazeProcessor) error {
 	r := text.NewReader(s)
 
 	// parse options are:
@@ -257,7 +258,7 @@ func splitByHeading(md goldmark.Markdown, s []byte, gp *cmds.GlazeProcessor) err
 // simpleLinearize is a simple walker that will linearize the blocks encountered,
 // and filter out the Document and Text blocks
 // to avoid duplicates, for a very simple document.
-func simpleLinearize(md goldmark.Markdown, s []byte, gp *cmds.GlazeProcessor) error {
+func simpleLinearize(md goldmark.Markdown, s []byte, gp *processor.GlazeProcessor) error {
 	r := text.NewReader(s)
 
 	// parse options are:
